requirements.txt
streamlit_lottie
streamlit-timeline

langchain # AI agent builder with many application
chainlit # fast gui to build chat gpt model
python-dotenv 
wikipedia  #langchain tools for searching wikipedia
duckduckgo-search #langchain tools for searching with duck duck go

plotly
pandas
numpy
scipy
scikit-learn

pyswarms